moved basic info code
