+++ src/app/child-dev-project/children/child-details/child-details.component.scss
@@ -24,51 +24,5 @@
   margin-top: 0;
 }
 
-.edit-button {
-  margin-left: 10px;
-}
-
-.mat-form-field {
-  margin-right: 10px;
-}
 
 
-.child-pic-container {
-  position: relative;
-}
-
-.child-pic {
-  width: 120px;
-  height: 150px;
-  object-fit: contain;
-  border: 1px solid black;
-}
-
-.child-pic-upload{
-  position: absolute;
-  left: 0;
-  width: 120px;
-  height: 150px;
-
-  border: none;
-  background: none;
-  cursor: pointer;
-}
-
-.upload-icon {
-  font-size: 2.0em;
-}
-
-.child-pic-photofile {
-  position: absolute;
-  left: 1px;
-  top: 115px;
-  width: 120px;
-  background: white;
-}
-
-.locked-icon {
-  line-height: 20px;
-  padding-left: 10px;
-  color: black;
-}
