+++ src/app/child-dev-project/children/child-details/child-details.component.ts
@@ -15,20 +15,15 @@
  *     along with ndb-core.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { Component, Inject, OnInit } from "@angular/core";
+import { Component, OnInit } from "@angular/core";
 import { Child } from "../model/child";
 import { EntityMapperService } from "../../../core/entity/entity-mapper.service";
-import { Gender } from "../model/Gender";
 import { ActivatedRoute, Router } from "@angular/router";
-import { FormBuilder, FormGroup, Validators } from "@angular/forms";
 import { MatSnackBar } from "@angular/material/snack-bar";
 import { Location } from "@angular/common";
 import { ConfirmationDialogService } from "../../../core/confirmation-dialog/confirmation-dialog.service";
 import * as uniqid from "uniqid";
-import { AlertService } from "../../../core/alerts/alert.service";
 import { ChildrenService } from "../children.service";
-import { ChildPhotoService } from "../child-photo-service/child-photo.service";
-import { SessionService } from "../../../core/session/session-service/session.service";
 import { UntilDestroy, untilDestroyed } from "@ngneat/until-destroy";
 
 @UntilDestroy()
@@ -39,154 +34,34 @@ import { UntilDestroy, untilDestroyed } from "@ngneat/until-destroy";
 })
 export class ChildDetailsComponent implements OnInit {
   child: Child = new Child("");
-
-  validateForm = false;
-  form: FormGroup;
   creatingNew = false;
-  editing = false;
-  enablePhotoUpload;
-  gender = Gender;
-
-  genders = Gender;
-  documentStatus = [
-    "OK (copy with us)",
-    "OK (copy needed for us)",
-    "needs correction",
-    "applied",
-    "doesn't have",
-    "not eligible",
-    "",
-  ];
-
-  isAdminUser: boolean;
 
   constructor(
     private entityMapperService: EntityMapperService,
     private childrenService: ChildrenService,
     private route: ActivatedRoute,
-    @Inject(FormBuilder) public fb: FormBuilder,
     private router: Router,
     private location: Location,
     private snackBar: MatSnackBar,
-    private confirmationDialog: ConfirmationDialogService,
-    private alertService: AlertService,
-    private childPhotoService: ChildPhotoService,
-    private sessionService: SessionService
+    private confirmationDialog: ConfirmationDialogService
   ) {
     this.route.paramMap.subscribe((params) => this.loadChild(params.get("id")));
-    this.isAdminUser = this.sessionService.getCurrentUser().admin;
-  }
-
-  initForm() {
-    this.form = this.fb.group({
-      name: [
-        { value: this.child.name, disabled: !this.editing },
-        Validators.required,
-      ],
-      // gender:         [{value: this.child.gender}], // reactive forms seem broken for mat-select, using ngModel instead
-      projectNumber: [
-        { value: this.child.projectNumber, disabled: !this.editing },
-      ],
-      dateOfBirth: [{ value: this.child.dateOfBirth, disabled: !this.editing }],
-      motherTongue: [
-        { value: this.child.motherTongue, disabled: !this.editing },
-      ],
-      religion: [{ value: this.child.religion, disabled: !this.editing }],
-
-      center: [
-        { value: this.child.center, disabled: !this.editing },
-        Validators.required,
-      ],
-      status: [{ value: this.child.status, disabled: !this.editing }],
-      admissionDate: [
-        { value: this.child.admissionDate, disabled: !this.editing },
-      ],
-
-      address: [{ value: this.child.address, disabled: !this.editing }],
-      phone: [{ value: this.child.phone, disabled: !this.editing }],
-      guardianName: [
-        { value: this.child.guardianName, disabled: !this.editing },
-      ],
-      preferredTimeForGuardianMeeting: [
-        {
-          value: this.child.preferredTimeForGuardianMeeting,
-          disabled: !this.editing,
-        },
-      ],
-
-      // aadhar:         [{value: this.child.has_aadhar,         disabled: !this.editing}],
-      // kanyashree:     [{value: this.child.has_kanyashree,     disabled: !this.editing}],
-      // bankAccount:    [{value: this.child.has_bankAccount,    disabled: !this.editing}],
-      // rationCard:     [{value: this.child.has_rationCard,     disabled: !this.editing}],
-      // bplCard:        [{value: this.child.has_BplCard,        disabled: !this.editing}],
-
-      photoFile: [{ value: this.child.photoFile, disabled: !this.editing }],
-    });
   }
 
   ngOnInit() {}
 
   loadChild(id: string) {
     if (id === "new") {
-      this.creatingNew = true;
-      this.editing = true;
       this.child = new Child(uniqid());
+      this.creatingNew = true;
     } else {
       this.childrenService
         .getChild(id)
         .pipe(untilDestroyed(this))
         .subscribe((child) => {
           this.child = child;
-          this.initForm();
         });
     }
-    this.initForm();
-  }
-
-  switchEdit() {
-    this.editing = !this.editing;
-    this.enablePhotoUpload = this.childPhotoService.canSetImage();
-    this.initForm();
-  }
-
-  save() {
-    // errors regarding invalid fields wont be displayed unless marked as touched
-    this.form.markAllAsTouched();
-    this.validateForm = true;
-
-    if (this.form.valid) {
-      this.assignFormValuesToChild(this.child, this.form);
-
-      this.entityMapperService
-        .save<Child>(this.child)
-        .then(() => {
-          if (this.creatingNew) {
-            this.router.navigate(["/child", this.child.getId()]);
-            this.creatingNew = false;
-          }
-          this.alertService.addInfo("Saving Successful");
-          this.switchEdit();
-        })
-        .catch((err) =>
-          this.alertService.addDanger(
-            'Could not save Child "' + this.child.name + '": ' + err
-          )
-        );
-    } else {
-      const invalidFields = this.getInvalidFields(this.form);
-      this.alertService.addDanger(
-        "Form invalid, required fields (" + invalidFields + ") missing"
-      );
-    }
-  }
-
-  private assignFormValuesToChild(child: Child, form: FormGroup) {
-    Object.keys(form.controls).forEach((key) => {
-      const value = form.get(key).value;
-      if (value !== null) {
-        child[key] = value;
-      }
-    });
   }
 
   removeChild() {
@@ -216,28 +91,5 @@ export class ChildDetailsComponent implements OnInit {
 
   navigateBack() {
     this.location.back();
-  }
-
-  getInvalidFields(form: FormGroup) {
-    const invalid = [];
-    const controls = this.form.controls;
-    for (const field in controls) {
-      if (controls[field].invalid) {
-        invalid.push(field);
-      }
-    }
-    return invalid;
-  }
-
-  /**
-   * hands over the selected file to the cloudFileService together with the childId
-   * @param event The event of the file upload dialog
-   */
-  async uploadChildPhoto(event) {
-    await this.childPhotoService.setImage(
-      event.target.files[0],
-      this.child.entityId
-    );
-    this.child.photo.next(await this.childPhotoService.getImage(this.child));
   }
 }
